chore(shared): move shallowCompare to object.ts
