+++ packages/shared/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './shallowCompare';
 export * from './number';
 export * from './spec';
 export * from './expression';
